수정: 중복 되는 함수  getUserIdFromToken 공통 함수로 묶음
